refactor: rearchitecture
