+++ lib/scanner/src/lib.rs
@@ -1,3 +1,1 @@
-pub mod buffer;
 pub mod scanner;
-mod token_generator;
